+++ src/ui/containers/ApplianceStatusContainer.java
@@ -6,9 +6,9 @@ import java.time.DayOfWeek;
  * Created by fegwin on 13/10/2016.
  */
 public class ApplianceStatusContainer extends StatusContainerBase {
-    public int consuming;
+    public double consuming;
 
-    public ApplianceStatusContainer(int consuming, int hourOfDay, DayOfWeek dayOfWeek) {
+    public ApplianceStatusContainer(double consuming, int hourOfDay, DayOfWeek dayOfWeek) {
         super(hourOfDay, dayOfWeek);
         this.consuming = consuming;
     }
